+++ Libraries/RobsJuceModules/romos/Modules/romos_FormulaModules.cpp
@@ -99,7 +99,7 @@ bool FormulaModule_1_1::setFormula(const std::string& newFormula)
 
 void FormulaModule_1_1::updateEvaluatorFormulas()
 {
-  for(int i = 0; i < evaluators.size(); i++)
+  for(size_t i = 0; i < evaluators.size(); i++)
   {
     evaluators[i]->initVariableList();  // without it, the evaluator may still remember old input
                                         // variables from previous formulas - maybe that can be
@@ -112,7 +112,7 @@ void FormulaModule_1_1::updateEvaluatorFormulas()
 void FormulaModule_1_1::updateInputVariables()
 {
   RAPT::rsAssert(inVariables.size() == evaluators.size());
-  for(int i = 0; i < evaluators.size(); i++)
+  for(size_t i = 0; i < evaluators.size(); i++)
     inVariables[i] = evaluators[i]->getVariableAddress("x");
 }
